refactor(auth): move login mutation hook to feature hooks directory
The login mutation hook was moved from the api directory to the feature hooks directory to better organize the codebase and follow feature-based architecture. The functionality remains unchanged but the error handling was slightly improved to use response.data.message instead of fullError.data.body.error.message.
